Setting up branch for Go only.
