--- packages/shrampy-dev/bot/requirements.txt
@@ -1,5 +1,0 @@
-urllib3<2
-Mastodon.py>=1.7.0
-twitchAPI==3.11.0
-hikari==2.1.0
-async-timeout
