--- packages/shrampy-prod/bot/lib/google.py
@@ -1,1 +1,0 @@
-# Future work in correlating a Google Sheet goes here
