+++ packages/shrampybot-dev/api/go.mod
@@ -1,0 +1,3 @@
+module main
+
+go 1.20
