+++ packages/shrampybot-dev/api/main.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"context"
+)
+
+func Main(ctx context.Context, event Event) Response {
+	resp := Response{
+		Body:       event.Http.Method,
+		StatusCode: "200",
+		Headers: ResponseHeaders{
+			ContentType: "application/json",
+		},
+	}
+	return resp
+}
