--- packages/shrampybot-dev/main/go.mod
@@ -1,5 +1,0 @@
-module main
-
-go 1.20
-
-require github.com/julienschmidt/httprouter v1.3.0 // indirect
